+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 v0.9.0 (2017-XX-XX)
 ...................
 * allow set encoding in msgpack for jobs #49
+* cron tasks allowing scheduling of functions in the future #50
 
 v0.8.1 (2017-06-05)
 ...................